Moved dc_load from odc.algo to odc.stac
The idea is to have minimal set of dependencies for a task

```
Given collection of STAC items
  make datacube compatible xr.Dataset
```
